requirements_web.txt
flask==3.0.0
gunicorn==21.2.0
pytesseract==0.3.10
Pillow==10.2.0
pypdf==4.0.1
reportlab==4.0.7
